correction et ajout des nouvelles routes

+++ index.php
@@ -12,7 +12,6 @@ require_once ROOT . 'controllers/Users.php';
 
 $security = new Users();
 define('USER', $security->authentificate());
-
 $router = new Router($_GET['url']);
 
 /** GET */
@@ -20,25 +19,27 @@ $router->get('/', "Bases#getHome");
 $router->get('/connexion', "Bases#getConnexion");
 $router->get('/activate', "Users#activate");
 
-/* if (USER) {
-} */
-$router->get('/tournois', "Tournaments#getHome");
-$router->get('/lesTournois', "Tournaments#getAllTournaments");
+if (USER !== false) {
+    $router->get('/tournois', "Tournaments#getHome");
+    $router->get('/lesTournois', "Tournaments#getAllTournaments");
+}
 
 // API
+$apiController = ["Users", "Tournaments"];
+
 $router->post('/ws/login', "Users#login");
 $router->post('/ws/users', "Users#post");
 
-$apiController = ["Users", "Tournaments"];
-foreach ($apiController as $controller) {
-    $router->get("/ws/$controller", "$controller#getAll");
-    $router->get("/ws/$controller/:id", "$controller#get");
-    $router->post("/ws/$controller/", "$controller#post");
-    $router->put("/ws/$controller/", "$controller#put");
-    $router->delete("/ws/$controller/:id", "$controller#delete");
+if (USER !== false) {
+    $router->post("/ws/tournaments/:id/join", "Tournaments#join");
+    foreach ($apiController as $controller) {
+        $router->get("/ws/$controller", "$controller#getAll");
+        $router->get("/ws/$controller/:id", "$controller#get");
+        $router->post("/ws/$controller/", "$controller#post");
+        $router->put("/ws/$controller/", "$controller#put");
+        $router->delete("/ws/$controller/:id", "$controller#delete");
+    }
 }
-/* if (USER) {
-} */
 
 try {
     $router->run();
